+++ backend/joboffer-service/src/main/java/com/joboffer/jobofferservice/model/JobOffer.java
@@ -34,5 +34,5 @@ public class JobOffer {
     private List<String> skills;
 
     // Reference to recruiter in another service
-    private Integer recruiterId;
+    private Double recruiterId;
 }
